# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telecom Service & Network Management System",
    "description": "Backend platform for managing telecom customers, plans, SIMs, devices, subscriptions, usage, network infrastructure, outages, support tickets, SLAs and operational analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/outages": {
      "post": {
        "tags": [
          "Network Outages"
        ],
        "summary": "Create Outage",
        "operationId": "create_outage_api_v1_outages_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Network Outages"
        ],
        "summary": "List Outages",
        "operationId": "list_outages_api_v1_outages_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outage_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OutageStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outage Status"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OutageSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number, starting at 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Column name to sort by",
              "title": "Sort By"
            },
            "description": "Column name to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Sort Order"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search",
              "title": "Search"
            },
            "description": "Free-text search"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_OutageOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outages/{outage_id}": {
      "get": {
        "tags": [
          "Network Outages"
        ],
        "summary": "Get Outage",
        "operationId": "get_outage_api_v1_outages__outage_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outages/{outage_id}/affected-customers": {
      "get": {
        "tags": [
          "Network Outages"
        ],
        "summary": "Affected Customers",
        "operationId": "affected_customers_api_v1_outages__outage_id__affected_customers_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffectedCustomerOut"
                  },
                  "title": "Response Affected Customers Api V1 Outages  Outage Id  Affected Customers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outages/{outage_id}/in-progress": {
      "post": {
        "tags": [
          "Network Outages"
        ],
        "summary": "Mark In Progress",
        "operationId": "mark_in_progress_api_v1_outages__outage_id__in_progress_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outages/{outage_id}/resolve": {
      "post": {
        "tags": [
          "Network Outages"
        ],
        "summary": "Resolve Outage",
        "operationId": "resolve_outage_api_v1_outages__outage_id__resolve_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outage Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutageResolve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AffectedCustomerOut": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "notified": {
            "type": "boolean",
            "title": "Notified"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "notified"
        ],
        "title": "AffectedCustomerOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutageCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "outage_type": {
            "$ref": "#/components/schemas/OutageType",
            "default": "unplanned"
          },
          "severity": {
            "$ref": "#/components/schemas/OutageSeverity"
          },
          "tower_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Tower Ids"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "expected_resolution": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Resolution"
          }
        },
        "type": "object",
        "required": [
          "title",
          "severity",
          "tower_ids",
          "start_time"
        ],
        "title": "OutageCreate"
      },
      "OutageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "outage_type": {
            "$ref": "#/components/schemas/OutageType"
          },
          "severity": {
            "$ref": "#/components/schemas/OutageSeverity"
          },
          "status": {
            "$ref": "#/components/schemas/OutageStatus"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "expected_resolution": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Resolution"
          },
          "actual_resolution": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Resolution"
          },
          "tower_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tower Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "outage_type",
          "severity",
          "status",
          "start_time",
          "expected_resolution",
          "actual_resolution"
        ],
        "title": "OutageOut"
      },
      "OutageResolve": {
        "properties": {
          "actual_resolution": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Resolution"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          }
        },
        "type": "object",
        "title": "OutageResolve"
      },
      "OutageSeverity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "OutageSeverity"
      },
      "OutageStatus": {
        "type": "string",
        "enum": [
          "open",
          "in_progress",
          "resolved"
        ],
        "title": "OutageStatus"
      },
      "OutageType": {
        "type": "string",
        "enum": [
          "planned",
          "unplanned"
        ],
        "title": "OutageType"
      },
...
      "PaginatedResponse_OutageOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OutageOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PaginatedResponse[OutageOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}